creating Queue class
