--- src/problem4/Element/Element.java
@@ -1,12 +1,0 @@
-package problem4.Element;
-
-public class Element<T> {
-    private T node;
-    private T preOrderSuccessor;
-    private Element next;
-
-    public Element(T node, T preOrderSuccessor) {
-        this.node = node;
-        this.preOrderSuccessor = preOrderSuccessor;
-    }
-}
